refactor(payer-logo): add src and types folders
