+++ packages/payer-logo/.npmignore
@@ -1,0 +1,1 @@
+**/*test.js
