+++ packages/payer-logo/package.json
@@ -4,6 +4,7 @@
   "author": "Evan Sharp <evan.sharp@availity.com>",
   "description": "Easy to use component to display the payer&#x27;s logo given the payer&#x27;s ID",
   "main": "index.js",
+  "types": "index.d.ts",
   "keywords": [
     "react",
     "availity"
@@ -27,16 +28,14 @@
   "devDependencies": {
     "@availity/api-axios": "^5.5.14",
     "@availity/api-core": "^7.0.2",
-    "@availity/localstorage-core": "^3.0.0",
     "axios": "^0.21.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
   "peerDependencies": {
     "@availity/api-axios": "^5.1.5",
-    "@availity/api-core": "^5.1.5",
-    "@availity/localstorage-core": "^2.6.1",
-    "axios": "^0.19.2",
+    "@availity/api-core": "^7.0.1",
+    "axios": "^0.21.1",
     "react": ">=16.8.3"
   }
 }
